lntemp+itest: remove unused code to prepare `lntemp`'s takeover
This commit removes the old code living in `lntest` to prepare
`lntemp`'s takeover.
